requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
Deprecated==1.2.14
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
mysqlclient==2.2.4
packaging==24.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.16.0
yarl==1.9.4
